Fix linter errors
