+++ book.rb
@@ -11,4 +11,16 @@ class Book
   def to_s
     "Title: \"#{@title}\", Author: #{author}"
   end
+
+  def to_json(*args)
+    {
+      JSON.create_id => self.class.name,
+      'title' => @title,
+      'author' => @author
+    }.to_json(*args)
+  end
+
+  def self.json_create(object)
+    new(object['title'], object['author'])
+  end
 end
