+++ books.json
@@ -1,0 +1,1 @@
+[{"json_class":"Book","title":"Root","author":"Author 1"},{"json_class":"Book","title":"Man","author":"C"}]
