+++ people.json
@@ -1,0 +1,1 @@
+[{"json_class":"Student","id":77,"name":"Tony","age":"13","parent_permission":false},{"json_class":"Student","id":724,"name":"Henry","age":"30","parent_permission":false}]
