+++ rental.rb
@@ -2,7 +2,7 @@ class Rental
   attr_accessor :date
   attr_reader :person, :book
 
-  def initialize(date, person, book)
+  def initialize(date, book, person)
     @date = date
 
     @person = person
@@ -14,5 +14,14 @@ class Rental
 
   def to_s
     "Date: #{@date}, Book: \"#{book.title}\" by #{book.author}"
+  end
+
+  def to_json(*args)
+    {
+      JSON.create_id => self.class.name,
+      'date' => @date,
+      'book_title' => @book.title,
+      'person_id' => @person.id
+    }.to_json(*args)
   end
 end
