+++ rentals.json
@@ -1,0 +1,1 @@
+[{"json_class":"Rental","date":"223","book_title":"Root","person_id":77},{"json_class":"Rental","date":"1","book_title":"Man","person_id":724}]
