+++ storage.rb
@@ -75,7 +75,7 @@ class Storage
     student
   end
 
-  def store(people:, rentals:, books:)
+  def persist(people:, rentals:, books:)
     File.open('people.json', 'w') { |f| f.write JSON.generate(people) } unless people.empty?
     File.open('books.json', 'w') { |f| f.write JSON.generate(books) } unless books.empty?
     File.open('rentals.json', 'w') { |f| f.write JSON.generate(rentals) } unless rentals.empty?
