+++ student.rb
@@ -20,4 +20,14 @@ class Student < Person
   def to_s
     "[Student] #{super}"
   end
+
+  def to_json(*args)
+    {
+      JSON.create_id => self.class.name,
+      'id' => id,
+      'name' => name,
+      'age' => age,
+      'parent_permission' => parent_permission
+    }.to_json(*args)
+  end
 end
