+++ teacher.rb
@@ -1,8 +1,8 @@
 require_relative 'person'
 
 class Teacher < Person
-  def initialize(age:, specialization:, name: 'Unknow', parent_permission: true)
-    super(name: name, age: age, parent_permission: parent_permission)
+  def initialize(age:, specialization:, name: 'Unknow')
+    super(name: name, age: age)
     @specialization = specialization
   end
 
@@ -12,5 +12,15 @@ class Teacher < Person
 
   def to_s
     "[Teacher] #{super}"
+  end
+
+  def to_json(*args)
+    {
+      JSON.create_id => self.class.name,
+      'id' => id,
+      'name' => name,
+      'age' => age,
+      'specialization' => @specialization
+    }.to_json(*args)
   end
 end
